att inventario e add info2
